pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Vasiliy Zuev <nemixvasya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.19"}
aiogram = "^3.0.0b7"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
alembic = "^1.11.1"
redis = "^4.6.0"
aioredis = "^2.0.1"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
